A simple function to mock code in pkg
'mock' changes value in a locked environment and invisibly returns the original value

'with_mock' calls mock with on-exit hook and executes an arbitrary code.

Finally 'with_mock_pkg' does the same as 'with_mock', but accepts a pkg name instead of environment. It changes the value in the package namespace as well as the attached environment

For packages, there is always a package namespace, and then once the package is attached, a copy of environment is created containing only exported objects. Both need to be mocked otherwise pkg::foo and foo would return different objects.